run workflows on PRs from external contributors
As we discussed in the last ideation, we would like to go ahead and
run the standard workflows even for external PRs.  They require
someone to click approve before they actually run, so this should
always be safe.

If we're reading the docs correctly, these should run without any
secrets as well, so we need to fix up tests and skip them if they need
secrets.  Opening this PR from a fork to give this a try and see what
happens.