add NewPluginSpec helper function
We want to be able to get plugins from arbitrary git repositories.  To
do this, we need to parse the URL we get passed, so we can resolve it
properly.

Introduce a new NewPluginSpec helper to do that, so we can use it
consistently throughout the codebase.

Also extend this method to recognize such plugin names, and deal with
them appropriately, setting the `PluginDownloadURL`, and specifying
that this is a remote plugin.

Note that we also need to mangle the names slightly, as `/` will
create new subfolders on the filesystem, which doesn't work well with
our current plugin detection mechanisms.  Rather than trying to make
that work, replace them with `_`, which currently cannot be used in
plugin names, and thus are safe to re-use here.